# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MarketPay microfinance and USSD payment platform for informal market vendors in Sierra Leone.",
        "title": "MarketPay API",
        "contact": {
            "name": "MarketPay Engineering",
            "email": "engineering@marketpay.sl"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/loans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "List loans by state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_pkg_pagination.Response-github_com_marketpay_backend_internal_loan_domain_model_Loan"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Apply for a loan",
                "parameters": [
                    {
                        "description": "Loan application",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_loan_interfaces_http.applyLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/loans/vendor/{vendor_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "List loans for a specific vendor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_pkg_pagination.Response-github_com_marketpay_backend_internal_loan_domain_model_Loan"
                        }
                    }
                }
            }
        },
        "/loans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get loan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/loans/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Approve a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review note",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_loan_interfaces_http.reviewLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/loans/{id}/disburse": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Disburse an approved loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Monime reference",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_loan_interfaces_http.disburseLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/loans/{id}/reject": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Reject a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_loan_interfaces_http.rejectLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/loans/{id}/schedule": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get repayment schedule for a loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.RepaymentSchedule"
                            }
                        }
                    }
                }
            }
        },
        "/repayments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repayments"
                ],
                "summary": "Make a loan repayment",
                "parameters": [
                    {
                        "description": "Repayment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_repayment_interfaces_http.repayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                        }
                    }
                }
            }
        },
        "/repayments/loans/{id}/default": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "repayments"
                ],
                "summary": "Mark a loan as defaulted",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_marketpay_backend_internal_loan_domain_model.FundingSource": {
            "type": "string",
            "enum": [
                "MFI_PARTNER",
                "NGO_FUND",
                "BANK_PARTNER"
            ],
            "x-enum-varnames": [
                "FundingSourceMFIPartner",
                "FundingSourceNGOFund",
                "FundingSourceBankPartner"
            ]
        },
        "github_com_marketpay_backend_internal_loan_domain_model.InterestType": {
            "type": "string",
            "enum": [
                "FLAT",
                "DECLINING_BALANCE"
            ],
            "x-enum-varnames": [
                "InterestTypeFlat",
                "InterestTypeDecliningBalance"
            ]
        },
        "github_com_marketpay_backend_internal_loan_domain_model.Loan": {
            "type": "object",
            "properties": {
                "commission_paid": {
                    "type": "boolean"
                },
                "commission_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "credit_score_at_time": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "disbursed_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.RepaymentFrequency"
                },
                "funded_by": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.FundingSource"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interest_rate": {
                    "type": "number"
                },
                "interest_type": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.InterestType"
                },
                "loan_type": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.LoanType"
                },
                "monime_reference": {
                    "type": "string"
                },
                "outstanding_amount": {
                    "type": "number"
                },
                "partner_id": {
                    "type": "string"
                },
                "principal_amount": {
                    "type": "number"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "review_note": {
                    "type": "string"
                },
                "reviewed_by": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.RepaymentSchedule"
                    }
                },
                "state": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.LoanState"
                },
                "term_weeks": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "vendor_id": {
                    "type": "string"
                }
            }
        },
        "github_com_marketpay_backend_internal_loan_domain_model.LoanState": {
            "type": "string",
            "enum": [
                "DRAFT",
                "PENDING_REVIEW",
                "AUTO_APPROVED",
                "UNDER_REVIEW",
                "APPROVED",
                "REJECTED",
                "DISBURSED",
                "ACTIVE",
                "CLOSED",
                "DEFAULTED"
            ],
            "x-enum-varnames": [
                "LoanStateDraft",
                "LoanStatePendingReview",
                "LoanStateAutoApproved",
                "LoanStateUnderReview",
                "LoanStateApproved",
                "LoanStateRejected",
                "LoanStateDisbursed",
                "LoanStateActive",
                "LoanStateClosed",
                "LoanStateDefaulted"
            ]
        },
        "github_com_marketpay_backend_internal_loan_domain_model.LoanType": {
            "type": "string",
            "enum": [
                "EMERGENCY_ADVANCE",
                "STARTER_LOAN",
                "GROWTH_LOAN"
            ],
            "x-enum-varnames": [
                "LoanTypeEmergencyAdvance",
                "LoanTypeStarterLoan",
                "LoanTypeGrowthLoan"
            ]
        },
        "github_com_marketpay_backend_internal_loan_domain_model.RepaymentFrequency": {
            "type": "string",
            "enum": [
                "BIWEEKLY",
                "MONTHLY"
            ],
            "x-enum-varnames": [
                "RepaymentFrequencyBiweekly",
                "RepaymentFrequencyMonthly"
            ]
        },
        "github_com_marketpay_backend_internal_loan_domain_model.RepaymentSchedule": {
            "type": "object",
            "properties": {
                "amount_paid": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installment_no": {
                    "type": "integer"
                },
                "interest_due": {
                    "type": "number"
                },
                "is_grace_period": {
                    "type": "boolean"
                },
                "loan_id": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "penalty_amount": {
                    "type": "number"
                },
                "principal_due": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_due": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_marketpay_backend_pkg_pagination.Response-github_com_marketpay_backend_internal_loan_domain_model_Loan": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.Loan"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_loan_interfaces_http.applyLoanRequest": {
            "type": "object",
            "required": [
                "amount",
                "frequency",
                "funded_by",
                "loan_type",
                "term_weeks"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.RepaymentFrequency"
                },
                "funded_by": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.FundingSource"
                },
                "group_id": {
                    "type": "string"
                },
                "loan_type": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_loan_domain_model.LoanType"
                },
                "partner_id": {
                    "type": "string"
                },
                "term_weeks": {
                    "type": "integer"
                }
            }
        },
        "internal_loan_interfaces_http.disburseLoanRequest": {
            "type": "object",
            "required": [
                "monime_reference"
            ],
            "properties": {
                "monime_reference": {
                    "type": "string"
                }
            }
        },
        "internal_loan_interfaces_http.rejectLoanRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_loan_interfaces_http.reviewLoanRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "type": "string"
                }
            }
        },
        "internal_repayment_interfaces_http.repayRequest": {
            "type": "object",
            "required": [
                "amount",
                "loan_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "loan_id": {
                    "type": "string"
                },
                "monime_reference": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}